Change int to Long Everything done and work Hello! I'm Spring MVC application 5.2.0 version by sep'19

+++ src/main/java/web/controller/CarsController.java
@@ -23,7 +23,7 @@ public class CarsController {
 	}
 
 	@GetMapping("/{id}")
-	public String show (@PathVariable("id") int id, Model model) {
+	public String show (@PathVariable("id") Long id, Model model) {
 		model.addAttribute("cars", carDao.show(id));
 		return "cars/show";
 	}
